+++ lib/element.ts
@@ -3,7 +3,6 @@ import {
 	print,
 } from '../util/string';
 import {setChildNodes} from '../util/debug';
-import {getCondition} from '../util/selector';
 import {AstNode} from './node';
 import {elementLike} from '../mixin/elementLike';
 import type {
@@ -27,14 +26,15 @@ import type {
 	ImageParameterToken,
 	TranscludeToken,
 } from '../internal';
-import type {TokenTypeMap, SelectedTokenTypes} from '../map';
 
 /* NOT FOR BROWSER */
 
 import fs from 'fs';
 import path from 'path';
 import {classes} from '../util/constants';
+import {getCondition} from '../util/selector';
 import {readOnly} from '../mixin/readOnly';
+import type {TokenTypeMap, SelectedTokenTypes} from '../map';
 
 declare type LinkTokens = LinkToken | RedirectTargetToken | ExtLinkToken | MagicLinkToken | ImageParameterToken;
 
@@ -169,26 +169,6 @@ export abstract class AstElement extends AstNode {
 
 		setChildNodes(this as AstElement as Token, i, 0, [node]);
 		return node;
-	}
-
-	/**
-	 * Get the closest ancestor node that matches the selector
-	 *
-	 * 最近的符合选择器的祖先节点
-	 * @param selector selector / 选择器
-	 */
-	closest<K extends SelectedTokenTypes>(selector: K): TokenTypeMap[K] | undefined;
-	closest<T = Token>(selector: string): T | undefined;
-	closest(selector: string): Token | undefined {
-		const condition = getCondition(selector, this);
-		let {parentNode} = this;
-		while (parentNode) {
-			if (condition(parentNode)) {
-				return parentNode;
-			}
-			({parentNode} = parentNode);
-		}
-		return undefined;
 	}
 
 	/** @private */